requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.1
imageio==2.36.1
kiwisolver==1.4.7
matplotlib==3.9.3
numpy==2.1.3
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
six==1.17.0
